requirements.txt
assertpy==0.14
psycopg2
SQLAlchemy==1.3.3
Routes==2.4.1
pymysql==0.9.3
pytest==4.5.0
pyyaml